test(stripe): extract hasstripemockfixtures trait, dedupe ~80 lines across two test files (review feedback)
Moves the subtle Stripe SDK statics capture-and-restore + per-class
Product/Plan creation from AccountSubscriptionTest and
SubscriptionsControllerTest into a shared trait at
tests/Concerns/HasStripeMockFixtures.php. The two consuming classes
keep their own setUp() (config block referencing the per-class plan IDs)
and their own test methods; everything else now lives in one place.

Net change: -83 lines across the three files (trait 146 new, 237 lines
removed from the two test files, 8 readded for the `use` declarations).
Filter run unchanged: Tests 22, Assertions 35, Failures 5 — the same
five pre-existing AccountSubscriptionTest stripe-mock semantic gaps
remain, no behavioural drift introduced by the extraction.
