+++ tests/Feature/AccountSubscriptionTest.php
@@ -2,46 +2,16 @@
 
 namespace Tests\Feature;
 
-use Stripe\Exception\ApiErrorException;
 use App\Exceptions\StripeException;
-use Stripe\Plan;
-use Stripe\Stripe;
-use Stripe\Product;
-use Tests\FeatureTestCase;
-use Illuminate\Support\Str;
-use Laravel\Cashier\Subscription;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
+use Laravel\Cashier\Subscription;
+use Tests\Concerns\HasStripeMockFixtures;
+use Tests\FeatureTestCase;
 
 class AccountSubscriptionTest extends FeatureTestCase
 {
     use DatabaseTransactions;
-
-    private const STRIPE_MOCK_KEY = 'sk_test_stripemockkey';
-    private const STRIPE_MOCK_BASE = 'http://stripe-mock:12111';
-
-    /**
-     * @var string
-     */
-    protected static $stripePrefix = 'cashier-test-';
-
-    /**
-     * @var string
-     */
-    protected static $productId;
-
-    /**
-     * @var string
-     */
-    protected static $monthlyPlanId;
-
-    /**
-     * @var string
-     */
-    protected static $annualPlanId;
-
-    private static ?string $originalApiBase = null;
-    private static ?string $originalApiKey = null;
-    private static ?string $originalApiVersion = null;
+    use HasStripeMockFixtures;
 
     public function setUp(): void
     {
@@ -57,92 +27,6 @@ class AccountSubscriptionTest extends FeatureTestCase
             'monica.paid_plan_annual_id' => static::$annualPlanId,
             'monica.paid_plan_annual_price' => 500,
         ]);
-    }
-
-    public static function setUpBeforeClass(): void
-    {
-        // phpunit.xml sets backupStaticProperties="false", so snapshot the
-        // Stripe SDK statics ourselves and restore them in tearDownAfterClass —
-        // otherwise this class leaks the stripe-mock pointer to anything else
-        // that touches \Stripe\Stripe directly later in the run.
-        self::$originalApiBase = Stripe::$apiBase;
-        self::$originalApiKey = Stripe::getApiKey();
-        self::$originalApiVersion = Stripe::getApiVersion();
-
-        Stripe::setApiKey(self::STRIPE_MOCK_KEY);
-        Stripe::$apiBase = env('STRIPE_API_BASE', self::STRIPE_MOCK_BASE);
-        // Pin a specific API version so stripe-mock's response shape stays deterministic
-        // across stripe-php upgrades. Cashier-mediated calls still use Cashier::STRIPE_VERSION
-        // internally — most of the test surface goes through that, not this pin.
-        Stripe::setApiVersion('2024-12-18.acacia');
-
-        static::$productId = static::$stripePrefix.'product-'.Str::random(10);
-        static::$monthlyPlanId = static::$stripePrefix.'monthly-'.Str::random(10);
-        static::$annualPlanId = static::$stripePrefix.'annual-'.Str::random(10);
-
-        Product::create([
-            'id' => static::$productId,
-            'name' => 'Monica Test Product',
-        ]);
-
-        Plan::create([
-            'id' => static::$monthlyPlanId,
-            'nickname' => 'Monthly',
-            'currency' => 'USD',
-            'interval' => 'month',
-            'billing_scheme' => 'per_unit',
-            'amount' => 100,
-            'product' => static::$productId,
-        ]);
-        Plan::create([
-            'id' => static::$annualPlanId,
-            'nickname' => 'Annual',
-            'currency' => 'USD',
-            'interval' => 'year',
-            'billing_scheme' => 'per_unit',
-            'amount' => 500,
-            'product' => static::$productId,
-        ]);
-    }
-
-    public static function tearDownAfterClass(): void
-    {
-        parent::tearDownAfterClass();
-
-        if (static::$monthlyPlanId) {
-            static::deleteStripeResource(new Plan(static::$monthlyPlanId));
-            static::$monthlyPlanId = null;
-        }
-        if (static::$annualPlanId) {
-            static::deleteStripeResource(new Plan(static::$annualPlanId));
-            static::$annualPlanId = null;
-        }
-        if (static::$productId) {
-            static::deleteStripeResource(new Product(static::$productId));
-            static::$productId = null;
-        }
-
-        // Restore the Stripe SDK statics we captured in setUpBeforeClass.
-        if (self::$originalApiBase !== null) {
-            Stripe::$apiBase = self::$originalApiBase;
-        }
-        if (self::$originalApiKey !== null) {
-            Stripe::setApiKey(self::$originalApiKey);
-        }
-        if (self::$originalApiVersion !== null) {
-            Stripe::setApiVersion(self::$originalApiVersion);
-        }
-    }
-
-    protected static function deleteStripeResource($resource)
-    {
-        try {
-            if (method_exists($resource, 'delete')) {
-                $resource->delete();
-            }
-        } catch (ApiErrorException $e) {
-            //
-        }
     }
 
     public function test_it_throw_an_error_on_subscribe()
